+++ src/attila/game_api.h
@@ -63,5 +63,6 @@ extern uintptr_t g_settlement_cb_initialize_addr;
 
 extern const uintptr_t OFFSET_MAX_UNITS_ARMY;
 extern const uintptr_t OFFSET_MAX_UNITS_NAVY;
+extern const uintptr_t OFFSET_MAX_TRAITS;
 
 void refresh_settlements_display();